Create README.md
This would similar to the all weather car mats except for a Car Seat. It would have the shape of the car seat but with catch trays on the side that flow to the bottom catch tray.  It would have a built in bottle or sippy cup holder.  It would have openings for the seat belt so that normal function of the car seat would not be comprised. It would clean up with a wipe or rinse with hose.